requirements.txt
openai>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
mcp>=1.0.0
httpx>=0.27.0 